server/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
python-multipart
requests
